requirements.txt
opencv-python
mediapipe
streamlit
numpy
Pillow
